refactor: update details and select components
