+++ src/react-components/SelectElement/SelectElement.tsx
@@ -1,17 +1,21 @@
-import { ChangeEvent, useState, useContext } from 'react';
-import { ISelectElements } from './types';
-import { AppContext } from '../../react-components/Contexts/AppContext';
+import { ChangeEvent } from 'react';
+import { useAppDispatch } from '../../store/hooks/redux';
+import { productsSlice } from '../../store/reducers/productsReducer';
 import './SelectElement.css';
 
 export default function SelectElement() {
-  const { setLimitPerPage } = useContext(AppContext);
-  const [options] = useState([10, 20, 30]);
+  const dispatch = useAppDispatch();
+
+  const handleSelectClick = (num: number) => {
+    dispatch(productsSlice.actions.setLimitPerPage(num));
+  };
+  const options = [10, 20, 30];
 
   return (
     <select
       className="select"
       onChange={(event: ChangeEvent<HTMLSelectElement>) =>
-        setLimitPerPage(parseInt(event.target.value, 10))
+        handleSelectClick(parseInt(event.target.value, 10))
       }
     >
       <option disabled>Select Items Quantity</option>
